ResidoPhython/helixcare-resido-backend-d1337c4839c2/requirements/common.txt
setuptools==80.9.0
wheel==0.43.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
amqp==5.2.0
# asgiref==3.4.1
billiard==4.2.1
boto3==1.18.36
botocore==1.21.40
celery==5.5.3
certifi==2024.07.04
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
decorator==5.1.0
Deprecated==1.2.13
dj-database-url==3.0.1
Django==4.2.16
django-celery-beat==2.8.1
django-concurrency==2.7
django-cors-headers==3.8.0
django-field-history==0.8.0
django-filter==25.1
django-oauth-toolkit==1.5.0
django-rq==2.4.1
django-simple-history==3.0.0
django-storages==1.14.6
django-tenants==3.9.0
django-timezone-field==5.0
django-waffle==5.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.7
drf-api-logger==1.1.10
django-rest-swagger==2.2.0
et-xmlfile==1.1.0
Faker==9.5.2
future==0.18.3
geocoder==1.38.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
# wheel==0.43.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.2.0
jwcrypto==1.5.6
kombu==5.5.2
MarkupSafe==2.0.1
nltk==3.9.1
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.0
packaging==22.0
pandas==2.3.3
prompt-toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.20
PyJWT==2.4.0
# pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2025.2
PyYAML==6.0.3
ratelim==0.1.6
redis==7.1.0
regex==2021.10.21
requests==2.32.2
rq==1.10.0
# ruamel.yaml==0.17.20
# ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.5.0
supervisor==4.2.2
text-unidecode==1.3
tqdm==4.66.3
uritemplate==4.1.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.5
wrapt==1.13.2
xmltodict==0.12.0
# Js2Py==0.74
azure-identity==1.16.1
azure-storage-blob==12.18.3
# livekit-server-sdk-python==1.0.0
django-log-request-id==2.1.0
xlrd==2.0.1
xhtml2pdf==0.2.16
weasyprint>=61.0
pydyf>=0.10.0
JSON-log-formatter==1.1.1
python-dotenv==1.0.1
pycryptodome==3.23.0
geopy==2.4.1
django_advanced_password_validation==1.2.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.32
py-moneyed==3.0
django-money==3.5.3
zipp==3.19.1
ua-parser==0.18.0
user-agents==2.2.0
pdfkit==1.0.0
py7zr==0.22.0
razorpay==1.4.2
ipython==8.37.0
livekit-api==0.8.0
nanoid==2.0.0
django-silk==5.3.2
python-keycloak==4.7.3
forex-python==1.8
kafka-python==2.1.5
dj-standard-enums==0.1.3
backend-core @ git+https://shivamsharma27:ATBBYKK4MUaBJWWVm4bdDapgQVYx44FA1354@bitbucket.org/helixcare/backend-core.git@v1.0.195
